Bug 2676: Added role-related stuff to API.
git-svn-id: https://svn.imcode.com/imcode/imcms/trunk@3912 bd66a97b-2aff-0310-9095-89ca5cabf5a6

+++ web/imcms/docs/apisamples/role_delete_role.jsp
@@ -1,5 +1,7 @@
 <%@ page import="com.imcode.imcms.api.*" errorPage="error.jsp" %>
 
+<html>
+<body>
 <H3>Delete a role named "Test role"</H3>
 
 <%
@@ -7,10 +9,17 @@
     UserService userService = imcmsSystem.getUserService();
 %>
 Before:<br>
-<%=java.util.Arrays.asList( userService.getAllRolesNames() )%>
+<%=java.util.Arrays.asList( userService.getAllRoles() )%><br>
 <%
-   String role = "Test role";
-   userService.deleteRole( role );
-%><br>
-After delete the role named "<%=role%>":<br>
-<%=java.util.Arrays.asList( userService.getAllRolesNames() )%>
+    String roleName = "Test role";
+    Role role = userService.getRole(roleName);
+    if (null != role) {
+        userService.deleteRole( role );
+        %>After delete of the role named "<%=role%>":<br>
+        <%=java.util.Arrays.asList( userService.getAllRoles() )%><%
+    } else {
+        %>Role "<%= roleName %>" does not exist.<%
+    }
+%>
+</body>
+</html>
